refactor(bible): extract prepared tab components to bible_prepared module
Splits bible.rs below the 1000-line hard cap by moving BiblePreparedTab,
PresentationCard, BiblePresentationModal, and BibleSettingsTab into their
own module, mirroring the existing bible_controls / bible_slides split.
