+++ crates/presenter-ui/src/pages/mod.rs
@@ -1,6 +1,7 @@
 pub mod ai;
 pub mod bible;
 pub mod bible_controls;
+pub mod bible_prepared;
 pub mod bible_slides;
 pub mod operator;
 pub mod settings;
